+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to
 [Compatible Versioning](https://github.com/staltz/comver).
 
 ## [Unreleased]
+
+## [0.3] - 2018-12-05
 ### Added
 - Encode/decode EpochTime
 - EDN-based schema DSL
@@ -47,6 +49,7 @@ and this project adheres to
   - RON-Schema
   - RON-Schema TemplateHaskell code generator
 
-[Unreleased]: https://github.com/ff-notes/ff/compare/v0.2...HEAD
+[Unreleased]: https://github.com/ff-notes/ff/compare/v0.3...HEAD
+[0.3]: https://github.com/ff-notes/ff/compare/v0.2...v0.3
 [0.2]: https://github.com/ff-notes/ff/compare/v0.1...v0.2
 [0.1]: https://github.com/ff-notes/ff/tree/v0.1